Yeay

+++ src/Interface_WGraph.java
@@ -1,7 +1,7 @@
 
 public interface Interface_WGraph {
 	
-	public void add_edge(int n1, int n2, int w);
+	public void add_edge(int n1, int n2, double w);
 	
 	public void remove_edge(int n1, int n2);
